requirements.txt
aiohttp == 3.8.5
discord.py == 2.3.2
requests == 2.31.0
google-generativeai == 0.3.1